Review Committing in ipv6calc Modified Files:  	Makefile ipv6calc.c showinfo.c showinfo.h

+++ ipv6calc/showinfo.h
@@ -1,13 +1,14 @@
 /*
  * Project    : ipv6calc
  * File       : showinfo.h
- * Version    : $Id: showinfo.h,v 1.3 2002/03/24 17:00:39 peter Exp $
+ * Version    : $Id: showinfo.h,v 1.4 2002/04/04 19:40:10 peter Exp $
  * Copyright  : 2001-2002 by Peter Bieringer <pb (at) bieringer.de>
  *
  * Information:
  *  Header file for showinfo.c
  */ 
 
+#include <stdint.h>
 #include "libipv6addr.h"
 #include "libipv4addr.h"
 #include "libmac.h"
@@ -15,7 +16,7 @@
 #define DEBUG_showinfo		0x2
 
 /* prototypes */
-extern int  showinfo_ipv6addr(const ipv6calc_ipv6addr *ipv6addrp, const unsigned long formatoptions);
-extern int  showinfo_ipv4addr(const ipv6calc_ipv4addr *ipv4addrp, const unsigned long formatoptions);
-extern int  showinfo_eui48(const ipv6calc_macaddr *macaddrp, const unsigned long formatoptions);
+extern int  showinfo_ipv6addr(const ipv6calc_ipv6addr *ipv6addrp, const uint32_t formatoptions);
+extern int  showinfo_ipv4addr(const ipv6calc_ipv4addr *ipv4addrp, const uint32_t formatoptions);
+extern int  showinfo_eui48(const ipv6calc_macaddr *macaddrp, const uint32_t formatoptions);
 extern void showinfo_availabletypes(void);
