+++ lib/librfc2874.h
@@ -1,17 +1,18 @@
 /*
  * Project    : ipv6calc
  * File       : librfc2874.h
- * Version    : $Id: librfc2874.h,v 1.1 2002/03/18 19:59:25 peter Exp $
+ * Version    : $Id: librfc2874.h,v 1.2 2002/04/04 19:40:27 peter Exp $
  * Copyright  : 2001-2002 by Peter Bieringer <pb (at) bieringer.de>
  *
  * Information:
  *  Header file for librfc2874.c
  */ 
 
+#include <stdint.h>
 #include "libipv6addr.h"
 
 /* prototypes */
-extern int librfc2874_addr_to_bitstring(const ipv6calc_ipv6addr *ipv6addrp, char *resultstring, const unsigned long int formatoptions);
+extern int librfc2874_addr_to_bitstring(const ipv6calc_ipv6addr *ipv6addrp, char *resultstring, const uint32_t formatoptions);
 extern int librfc2874_bitstring_to_ipv6addrstruct(const char *inputstring, ipv6calc_ipv6addr *ipv6addrp, char *resultstring);
 
 extern int librfc2874_formatcheck(const char *string, char *infostring);
